+++ apps/desktop/src-tauri/src/features/checkout/models/checkout_model.rs
@@ -15,11 +15,11 @@ pub struct Checkout {
     pub shipping_line: Option<String>,          // JSONB stored as TEXT
     pub applied_discount_codes: Option<String>, // JSONB stored as TEXT
     pub currency: Option<String>,               // DEFAULT 'BRL'
-    pub subtotal_price: Option<f64>,            // REAL in SQLite
-    pub total_tax: Option<f64>,
-    pub total_shipping: Option<f64>,
-    pub total_discounts: Option<f64>,
-    pub total_price: Option<f64>,
+    pub subtotal_price: Option<i64>,  // centavos
+    pub total_tax: Option<i64>,       // centavos
+    pub total_shipping: Option<i64>,  // centavos
+    pub total_discounts: Option<i64>, // centavos
+    pub total_price: Option<i64>,     // centavos
     pub status: Option<String>, // DEFAULT 'open'
     pub reservation_expires_at: Option<String>,
     pub completed_at: Option<String>,
