+++ apps/desktop/src-tauri/src/features/pos_session/models/pos_session_model.rs
@@ -13,26 +13,26 @@ pub struct PosSession {
     pub status: Option<String>, // 'open', 'paused', 'closed', 'cancelled'
 
     // Opening values
-    pub opening_cash_amount: Option<f64>,
+    pub opening_cash_amount: Option<i64>, // centavos
     pub opening_notes: Option<String>,
     pub opened_at: Option<String>,
 
     // Closing values
-    pub closing_cash_amount: Option<f64>,
+    pub closing_cash_amount: Option<i64>, // centavos
     pub closing_notes: Option<String>,
     pub closed_at: Option<String>,
     pub closed_by: Option<String>,
 
     // Session totals
-    pub total_sales: Option<f64>,
-    pub total_returns: Option<f64>,
-    pub total_cash_in: Option<f64>,
-    pub total_cash_out: Option<f64>,
+    pub total_sales: Option<i64>,    // centavos
+    pub total_returns: Option<i64>,  // centavos
+    pub total_cash_in: Option<i64>,  // centavos
+    pub total_cash_out: Option<i64>, // centavos
     pub transaction_count: Option<i32>,
 
     // Cash difference
-    pub expected_cash_amount: Option<f64>,
-    pub cash_difference: Option<f64>,
+    pub expected_cash_amount: Option<i64>, // centavos
+    pub cash_difference: Option<i64>,      // centavos
 
     pub metadata: Option<String>,
     #[serde(rename = "_status")]
